Moved the errors into their own module
